refactor(editor): Migrate versions api to rest-client-api
