+++ packages/@n8n/constants/src/index.ts
@@ -1,4 +1,5 @@
 export * from './browser';
+export * from './instance';
 
 export const LICENSE_FEATURES = {
 	SHARING: 'feat:sharing',
@@ -96,9 +97,3 @@ export const LDAP_DEFAULT_CONFIGURATION: LdapConfig = {
 	searchPageSize: 0,
 	searchTimeout: 60,
 };
-
-export const INSTANCE_TYPES = ['main', 'webhook', 'worker'] as const;
-export type InstanceType = (typeof INSTANCE_TYPES)[number];
-
-export const INSTANCE_ROLES = ['unset', 'leader', 'follower'] as const;
-export type InstanceRole = (typeof INSTANCE_ROLES)[number];
