+++ packages/frontend/@n8n/rest-api-client/src/api/index.ts
@@ -9,4 +9,5 @@ export * from './npsSurvey';
 export * from './orchestration';
 export * from './roles';
 export * from './ui';
+export * from './versions';
 export * from './webhooks';
