+++ packages/frontend/@n8n/rest-api-client/src/api/versions.ts
@@ -1,0 +1,39 @@
+import { INSTANCE_ID_HEADER } from '@n8n/constants';
+import type { INodeParameters } from 'n8n-workflow';
+
+import { get } from '../utils';
+
+export interface VersionNode {
+	name: string;
+	displayName: string;
+	icon: string;
+	iconUrl?: string;
+	defaults: INodeParameters;
+	iconData: {
+		type: string;
+		icon?: string;
+		fileBuffer?: string;
+	};
+	typeVersion?: number;
+}
+
+export interface Version {
+	name: string;
+	nodes: VersionNode[];
+	createdAt: string;
+	description: string;
+	documentationUrl: string;
+	hasBreakingChange: boolean;
+	hasSecurityFix: boolean;
+	hasSecurityIssue: boolean;
+	securityIssueFixVersion: string;
+}
+
+export async function getNextVersions(
+	endpoint: string,
+	version: string,
+	instanceId: string,
+): Promise<Version[]> {
+	const headers = { [INSTANCE_ID_HEADER as string]: instanceId };
+	return await get(endpoint, version, {}, headers);
+}
