+++ packages/frontend/editor-ui/src/Interface.ts
@@ -45,6 +45,7 @@ import type {
 	ITaskData,
 	ISourceData,
 } from 'n8n-workflow';
+import type { VersionNode } from '@n8n/rest-api-client/api/versions';
 
 import type {
 	AI_NODE_CREATOR_VIEW,
@@ -900,33 +901,7 @@ export interface ITagRow {
 	canDelete?: boolean;
 }
 
-export interface IVersion {
-	name: string;
-	nodes: IVersionNode[];
-	createdAt: string;
-	description: string;
-	documentationUrl: string;
-	hasBreakingChange: boolean;
-	hasSecurityFix: boolean;
-	hasSecurityIssue: boolean;
-	securityIssueFixVersion: string;
-}
-
-export interface IVersionNode {
-	name: string;
-	displayName: string;
-	icon: string;
-	iconUrl?: string;
-	defaults: INodeParameters;
-	iconData: {
-		type: string;
-		icon?: string;
-		fileBuffer?: string;
-	};
-	typeVersion?: number;
-}
-
-export interface ITemplatesNode extends IVersionNode {
+export interface ITemplatesNode extends VersionNode {
 	id: number;
 	categories?: ITemplatesCategory[];
 }
